requirements.txt
numpy==1.22.3
pandas==1.4.2
psycopg2-binary==2.9.3
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
